udisksctl: fix paths with a colon
For paths such as `/dev/disk/by-id/usb-JetFlash_Transcend_64GB_15AFGA849642-0:0`, `_describe` parses the colon as field separator. This results in wrong completions and the general feeling that udisksctl is broken because it won't mount my pendrive.

Since the output of `udisksctl complete` is (apparently) just a list of paths, it's safe to escape all colons with a backslash.